+++ src/senaite/core/browser/controlpanel/analysisprofiles/view.py
@@ -23,79 +23,89 @@ import collections
 from bika.lims import api
 from bika.lims import bikaMessageFactory as _
 from bika.lims.utils import get_link_for
-from senaite.app.listing import ListingView
+from senaite.app.listing.view import ListingView
 from senaite.core.catalog import SETUP_CATALOG
 from senaite.core.i18n import translate
-from senaite.core.permissions import AddPreservation
+from senaite.core.permissions import AddAnalysisProfile
 
 
-class SamplePreservationsView(ListingView):
+class AnalysisProfilesView(ListingView):
+    """Controlpanel Listing for Analysis Profiles
+    """
 
     def __init__(self, context, request):
-        super(SamplePreservationsView, self).__init__(context, request)
+        super(AnalysisProfilesView, self).__init__(context, request)
 
         self.catalog = SETUP_CATALOG
         self.show_select_column = True
 
         self.contentFilter = {
-            "portal_type": "SamplePreservation",
+            "portal_type": "AnalysisProfile",
             "sort_on": "sortable_title",
             "sort_order": "ascending",
         }
 
         self.context_actions = {
-            _(u"listing_samplepreservations_action_add", default=u"Add"): {
-                "url": "++add++SamplePreservation",
-                "permission": AddPreservation,
+            _(u"listing_analysisprofiles_action_add", default=u"Add"): {
+                "url": "++add++AnalysisProfile",
+                "permission": AddAnalysisProfile,
                 "icon": "senaite_theme/icon/plus"
             }
         }
 
         self.title = translate(_(
-            u"listing_samplepreservations_title",
-            default=u"Sample Preservations")
+            u"listing_analysisprofiles_title",
+            default=u"Analysis Profiles")
         )
-        self.icon = api.get_icon("SamplePreservations", html_tag=False)
+        self.icon = api.get_icon("AnalysisProfiles", html_tag=False)
 
         self.columns = collections.OrderedDict((
             ("Title", {
                 "title": _(
-                    u"listing_samplepreservations_column_title",
-                    default=u"Sample Preservation"
+                    u"listing_analysisprofiles_column_title",
+                    default=u"Profile"
                 ),
-                "index": "sortable_title"}),
+                "index": "sortable_title",
+            }),
             ("Description", {
                 "title": _(
-                    u"listing_samplepreservations_column_description",
+                    u"listing_analysisprofiles_column_description",
                     default=u"Description"
                 ),
                 "index": "Description",
-                "toggle": True}),
+                "toggle": True,
+            }),
+            ("ProfileKey", {
+                "title": _(
+                    u"listing_analysisprofiles_column_profilekey",
+                    default=u"Profile Key"
+                ),
+                "sortable": False,
+                "toggle": True,
+            }),
         ))
 
         self.review_states = [
             {
                 "id": "default",
                 "title": _(
-                    u"listing_samplepreservations_state_active",
+                    u"listing_analysisprofiles_state_active",
                     default=u"Active"
                 ),
                 "contentFilter": {"is_active": True},
-                "transitions": [{"id": "deactivate"}, ],
                 "columns": self.columns.keys(),
             }, {
                 "id": "inactive",
                 "title": _(
-                    u"listing_samplepreservations_state_inactive",
+                    u"listing_analysisprofiles_state_inactive",
                     default=u"Inactive"
                 ),
                 "contentFilter": {'is_active': False},
-                "transitions": [{"id": "activate"}, ],
                 "columns": self.columns.keys(),
             }, {
                 "id": "all",
                 "title": _(
-                    u"listing_samplepreservations_state_all",
+                    u"listing_analysisprofiles_state_all",
                     default=u"All"
                 ),
                 "contentFilter": {},
@@ -104,5 +114,7 @@ class SamplePreservationsView(ListingView):
         ]
 
     def folderitem(self, obj, item, index):
+        obj = api.get_object(obj)
         item["replace"]["Title"] = get_link_for(obj)
+        item["ProfileKey"] = obj.getProfileKey()
         return item