+++ src/senaite/core/interfaces/__init__.py
@@ -163,6 +163,16 @@ class IHaveUIDReferences(Interface):
     """
 
 
+class IAnalysisProfile(Interface):
+    """Marker interface for analysis profiles
+    """
+
+
+class IAnalysisProfiles(Interface):
+    """Marker interface for analysis profiles setup folder
+    """
+
+
 class ISampleContainers(Interface):
     """Marker interface for sample container setup folder
     """
@@ -252,4 +262,19 @@ class IASTMImporter(Interface):
 
     def import_data(data):
         """Import the processed JSON data from the wrapper
+        """
+
+
+class IClientAwareMixin(Interface):
+    """Marker interface for objects that can be bound to a Client, either
+    because they can be added inside a Client folder or because it can be
+    assigned through a Reference field
+    """
+
+    def getClient(self):
+        """Returns the client this object is bound to, if any
+        """
+
+    def getClientUID(self):
+        """Returns the client UID this object is bound to, if any
         """