+++ src/senaite/core/setuphandlers.py
@@ -246,6 +246,7 @@ def add_senaite_setup_items(portal):
         ("departments", "Lab Departments", "Departments"),
         ("labels", "Labels", "Labels"),
         ("sampleconditions", "Sample Conditions", "SampleConditions"),
+        ("analysisprofiles", "Analysis Profiles", "AnalysisProfiles"),
         ("samplepreservations", "Sample Preservations", "SamplePreservations"),
     ]
     setup = api.get_senaite_setup()